Implemented most of the program with the specified flags.
Helper now has the ability to set wallpaper to next, previous, and to initialize from a file.
This file is supposed to write to .config/wallpaper-helper/wallpaper, as it is
a user-specified configuration.
I have added some functions to make it a bit easier to use different
wallpaper utilities than hyprpaper, and these will be added as soon as the
program is complete.
Next commit will rename the repository to wallpaper-helper, as a wallpaper
utility is already a different thing.